(edit) delete helpers/search directory and move all the helpers to helpers/ and adjust imports in codebase
