+++ src/helpers/isFilterChecked.js
