--- src/helpers/search/index.js
@@ -1,4 +1,0 @@
-export { default as getAccessionCodesParam } from './getAccessionCodesParam'
-export { default as getFilterParam } from './getFilterParam'
-export { default as getQueryParam } from './getQueryParam'
-export { default as isChecked } from './isChecked'
